+++ src/components/Main/Selector/SelectorForm/SelectorForm.styles.jsx
@@ -1,4 +1,5 @@
 import { styled } from "styled-components";
+import { appred } from "../../../../utils/colors.styles";
 
 export const $SelectorForm = styled.form`
   margin: 25px 100px 25px 0;
@@ -6,6 +7,12 @@ export const $SelectorForm = styled.form`
   justify-content: flex-start;
   column-gap: 10px;
   align-items: center;
+`
+
+export const $ErrorLabel = styled.div`
+  color: ${appred};
+  text-align: center;
+  font-size: 14px;
 `
 
 export default $SelectorForm;